split backend into components
