--- source/component/component.py
@@ -1,2 +1,0 @@
-def add(num1, num2):
-    return num1 + num2
